Fix ruRU Localisation for private servers.
Why ? Because private server combat log is not fully localized. Therefore, we were not seeing correct icons and timers ( in fact they didn't work at all).
IF on your server totem timers and icons do not work properly,
Go and download CurseForge 8.1d version and see if it works...
Here is how combat log looks for me and had to be fixed:
 https://cdn.discordapp.com/attachments/766767499987386402/1131782235419324527/Wow_VOVOYKtvYs.png